Server(AI)/requirements.txt
fastapi
pydantic
ollama
openai-whisper
logging
uvicorn
python-multipart
